requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.10           # via django
django-extensions==3.0.4  # via -r requirements.in
django==3.1               # via -r requirements.in, djangorestframework
djangorestframework==3.11.1  # via -r requirements.in
inky==0.0.5               # via -r requirements.in
numpy==1.19.1             # via inky
pillow==7.2.0             # via -r requirements.in
pytz==2020.1              # via django
rpi.gpio==0.7.0           # via inky
smbus2==0.3.0             # via inky
spidev==3.5               # via inky
sqlparse==0.3.1           # via django
